# doc/a3s-openapi.yaml
openapi: 3.0.0
info:
  description: >-
    API Definition for the A3S. This service allows authentication, authorisation and accounting.
  version: 1.0.1
  title: A3S
paths:
...
  '/roles':
    get:
      tags:
        - Role
      description: Search for Roles.
      summary: Search for Roles.
      operationId: listRoles
      parameters:
        - name: users
          in: query
          required: false
          description: Whether to fill in the users member field
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          required: false
          description: The page to view.
          schema:
            type: integer
            minimum: 0
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          description: The size of a page.
          schema:
            type: integer
            minimum: 1
            maximum: 20
            format: int32
            default: 10
        - name: filterDescription
          in: query
          required: false
          description: A search query filter on the role's description.
          schema:
            type: string
            minLength: 0
            maxLength: 255
          example: FilterByEntityType
        - name: orderBy
          in: query
          required: false
          style: form
          explode: false
          description: >-
            a comma separated list of fields in their sort order. Ascending
            order is assumed. Append desc after a field to indicate descending
            order.
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK.
          headers:
            X-Pagination:
              description: Pagination header.
              schema:
                $ref: '#/components/schemas/PaginationHeaderResponse'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the list of roles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Roles list not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Role
      summary: Create a Role.
      description: Create a new Role.
      operationId: createRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleSubmit'
      responses:
        '200':
          description: Successful. Role created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleSubmit'
        '400':
          description: Invalid parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Not authorized to create a Role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role related entity not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Non-Processible Entity. The request was correctly structured, but some business rules were violated, preventing the creation of the role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/roles/{roleId}':
    get:
      tags:
        - Role
      summary: Get a role.
      description: Get a role by its UUID.
      operationId: getRole
      parameters:
        - name: roleId
          in: path
          description: The UUID of the role.
          required: true
          schema:
            $ref: '#/components/schemas/RoleId'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access this role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Role
      summary: Update a role.
      description: Update a role by its UUID.
      operationId: updateRole
      parameters:
        - name: roleId
          in: path
          description: The UUID of the role.
          required: true
          schema:
            $ref: '#/components/schemas/RoleId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleSubmit'
      responses:
        '200':
          description: OK.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to update a role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role related entity not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Non-Processible Entity. The request was correctly structured, but some business rules were violated, preventing the updating of the role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  #endregion - Role
  #region - Teams
...
  '/permissions':
    get:
      tags:
        - Permission
      description: Search for permissions.
      summary: Search for permissions.
      operationId: listPermissions
      parameters:
        - name: page
          in: query
          required: false
          description: The page to view.
          schema:
            type: integer
            minimum: 0
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          description: The size of a page.
          schema:
            type: integer
            minimum: 1
            maximum: 20
            format: int32
            default: 10
        - name: filterDescription
          in: query
          required: false
          description: A search query filter on the description
          schema:
            type: string
            minLength: 0
            maxLength: 255
          example: FilterByEntityType
        - name: orderBy
          in: query
          required: false
          style: form
          explode: false
          description: >-
            a comma separated list of fields in their sort order. Ascending
            order is assumed. Append desc after a field to indicate descending
            order.
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          headers:
            X-Pagination:
              description: Pagination header.
              schema:
                $ref: '#/components/schemas/PaginationHeaderResponse'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the list of permissions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Permission list not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/permissions/{permissionId}':
    get:
      tags:
        - Permission
      summary: Get a permission.
      description: Get a permission by its UUID.
      operationId: getPermission
      parameters:
        - name: permissionId
          in: path
          description: permission
          required: true
          schema:
            $ref: '#/components/schemas/PermissionId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permission'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - You are not authorized to access the permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Permission not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        500:
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  #endregion
  
  
  # Region - Functions
...
components:
  schemas:
...
    RoleId:
      type: string
      format: uuid
    Role:
      description: |
        A Role within A3S. Roles can have functions assigned to them. Roles can be assigned to users.
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/RoleId'
        name:
          type: string
        description:
          type: string
        functionIds:
          type: array
          items:
            description: The UUIDS of the functions assigned to this role.
            type: string
            format: uuid
        userIds:
          description: The UUIDs of the users attached to the role
          type: array
          items:
            type: string
            format: uuid
        roleIds:
          description: The UUIDs of the child roles attached to the role.
          type: array
          items:
            $ref: '#/components/schemas/RoleId'
    RoleSubmit:
      description: |
        Model used to create or update a Role.
      required:
        - uuid
        - name
        - description
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        uuid:
          type: string
          format: uuid
        roleIds:
          description: The UUIDs of the child roles attached to the role.
          type: array
          items:
            $ref: '#/components/schemas/RoleId'
        functionIds:
          type: array
          items:
            description: The UUIDS of the functions assigned to this role.
            type: string
            format: uuid
        
    #endregion

    #region - Application
...
    PermissionId:
      type: string
      format: uuid
    Permission:
      description: |
        Permission for an application
      type: object
      properties:
        uuid:
          $ref: '#/components/schemas/PermissionId'
        name:
          type: string
        description:
          type: string

    # End Region - Permissions      
          
    # Region - Functions
...
    ErrorResponse:
      description: Error response object to give more info with HTTP status code
      type: object
      required:
        - message
      properties:
        message:
          type: string
    PaginationHeaderResponse:
      description: >-
        This is the standard response for requests that have paginated
        collections
      type: object
      properties:
        total:
          type: integer
          format: int32
          minimum: 1
          description: The total number of pages in the result set
        count:
          type: integer
          format: int32
          minimum: 1
          description: The total number of results in the result set
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 20
          default: 10
          description: >-
            The number items to include in a page of results. The page size for
            the result set
        current:
          type: integer
          format: int32
          minimum: 1
          description: >-
            The position of the page in the paged result set that is being
            returned
        first:
          type: string
          format: url
          description: >-
            The link to the first page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters.
        last:
          type: string
          format: url
          description: >-
            The link to the last page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters. If this is the first page in the result set then this
            will be [null].
        prev:
          type: string
          format: url
          description: >-
            The link to the previous page of results containing [size] results.
            This link includes the page, size, filter, orderBy and fields query
            parameters. If this is the first page in the result set then this
            will be [null].
        next:
          type: string
          format: url
          description: >-
            The link to the next page of results containing [size] results. This
            link includes the page, size, filter, orderBy and fields query
            parameters. If this is the last page in the result set then this will
            be [null].


    #endregion